test(work-graph): one fake subtree walk, not two (sage on #578)
Two test fakes had each grown their own copy of the `readSubtree` walk —
pre-order, `seen` guard, parent-from-the-edge — which is the same "two answers"
problem the production code just finished removing from `NodeState` assembly,
one layer down. Two copies of a seam's semantics drift, and a drifted fake makes
the contract layer's tests agree with a backend nobody ships.

`walkFakeSubtree` in `test/fixtures/` is the reference walk; a fake now supplies
only what makes it a fake, its child lookup and its node read.
